requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain_openai
twscrape
python-dotenv 
langfuse
streamlit
plotly
pytest
pytest-asyncio
pytest-mock
httpx